Refactor binlog and use in dotnet test and dotnet run
